backend/requirements.txt
asgiref==3.10.0
Django==4.2.25
django-cors-headers==4.9.0
django-prometheus==2.4.1
djangorestframework==3.16.1
pillow==11.3.0
prometheus_client==0.23.1
psycopg2-binary==2.9.11
sqlparse==0.5.3
typing_extensions==4.15.0
python-logging-loki